requirements.txt
langchain>=0.2.14
langchain-openai>=0.1.7
langchain-google-genai>=1.0.7
google-generativeai>=0.7.2
pydantic>=2.8.2
pypdf>=5.0.1
python-dotenv>=1.0.1
rich>=13.7.1
groq